server/requirements.txt
attrs==20.3.0
bravado==11.0.2
bravado-core==5.17.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.5
msgpack==1.0.0
neptune-client==0.4.126
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
py3nvml==0.2.6
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
typing-extensions==3.7.4.3
urllib3==1.26.2
webcolors==1.11.1
websocket-client==0.57.0
Werkzeug==1.0.1
xmltodict==0.12.0
zipp==3.4.0
